+++ end_to_end_tests/resources/page_locators/overview.robot
@@ -1,2 +1,3 @@
 ***Variables***
 ${timeline_locator}    xpath://*[@id="timeLineContainer"]
+${siteHeader}    xpath://*[@id="siteHeader"]
