+++ frontend/src/components/overview/Series.js
@@ -1,5 +1,8 @@
-﻿import React from 'react';
+﻿/* eslint-disable react-hooks/exhaustive-deps */
+import React, { useEffect } from 'react';
 import SuiteInstability from '../graphs/SuiteInstability';
+import { useParams } from 'react-router';
+import { useStateValue } from '../../contexts/state';
 import TimeLineChart from '../graphs/TimeLineChart';
 import DashboardList from '../dashlist/ListMain';
 import LastBuildElement from '../lastBuildElement/LastBuild'
@@ -22,6 +25,49 @@ const ElementHeader = styled.h3`
 
 const Series = () => {
     const [t] = useTranslation(['overview']);
+
+    const { seriesId } = useParams();
+
+    const [
+        {
+            parentData: { seriesData, buildData },
+            branchesState,
+        },
+        dispatch,
+    ] = useStateValue();
+
+    useEffect(() => {
+        if (branchesState) {
+            const branch = branchesState.series?.find(
+                ({ id: serie_id }) => serie_id === parseInt(seriesId, 10)
+            );
+            const fetchData = async () => {
+                dispatch({ type: 'setSeriesData', seriesData: branch });
+                dispatch({
+                    type: 'setSelectedBranch',
+                    name: branch.name || ' ',
+                    id: seriesId,
+                    team: branch.team || ' ',
+                });
+                try {
+                    const { last_build } = branch;
+                    const buildUrl = `/data/series/${seriesId}/builds/${last_build}/info?`;
+                    const res = await fetch(buildUrl);
+                    const json = await res.json();
+                    const buildData = json.build;
+
+                    dispatch({ type: 'setBuildData', buildData });
+                } catch (error) {
+                    dispatch({ type: 'setErrorState', errorState: error });
+                }
+            };
+
+            fetchData();
+        }
+        return () => {
+            dispatch({ type: 'flushParentData' });
+        };
+    }, [seriesId, branchesState]);
 
     return (
         <>
